sfotipy/requirements.txt
decorator==4.0.2
Django==1.8.4
django-extensions==1.5.6
django-grappelli==2.7.1
django-mockups==0.4.8
django-redis-cache==1.6.0
djangorestframework==3.2.3
hiredis==0.2.0
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
redis==2.10.3
simplegeneric==0.8.1
six==1.9.0
sorl-thumbnail==12.3
tablib==0.10.0
traitlets==4.0.0
wheel==0.24.0
